feat(chrono): refactor chrono support and fuzz testing
- Separate chrono and time feature support by moving chrono-specific code into dedicated modules/files and removing the shared Date component
- Implement conversion from DateTimeDecode to chrono::NaiveDateTime and remove the previous conversion from DateTimeEncode
- Remove support for chrono::DateTime<FixedOffset>
- Add fuzz testing for:
  - chrono::NaiveTime
  - chrono::NaiveDate
  - chrono::NaiveDateTime
  - chrono::DateTime<Utc>

This change improves modularity, eliminates unnecessary shared types, and strengthens testing coverage for the remaining chrono types.
