--- src/ext/date/chrono.rs
@@ -1,6 +1,0 @@
-#[cfg(feature = "chrono_datetime_fixedoffset")]
-mod date_time_fixed_offset;
-mod date_time_utc;
-mod naive_date;
-mod naive_date_time;
-mod naive_time;
