+++ src/ext/mod.rs
@@ -7,6 +7,8 @@ mod glam;
 mod chrono;
 #[cfg(feature = "rust_decimal")]
 mod rust_decimal;
+#[cfg(feature = "time")]
+mod time;
 #[cfg(feature = "uuid")]
 mod uuid;
